Backend/requirements.txt
geopy
flatlib
google-generativeai
python-dotenv
fastapi
uvicorn[standard]